DetailView major update
Added two strings for the detailview to use as a title for the
information and as an error in case nothing is found.
Changed the default artist image in detail_view.xml to our logo
Added two textViews to detail_view.xml to accommodate a complete address
Removed two obsolete TextViews from detail_view.xml

Changed all the == comparisons in OpenWeatherApi to .equalsIgnoreCase()
as they did not compare properly

All the TextViews in DetailView now update properly
The Ticket button works as intended, in case no ticket url is provided
we generate a google URL
Ratings update properly, however they do not show on the RatingBars yet.
No clue why
Artist image updates when we have an image url available. Unfortunately,
sometimes we actually get a blank string as a returned image URL, so
every now and then it will fail and display a blank image.