Error pop-ups and search from different activities
When a fatal error occurs and no events are returned the user now gets
to see a pop-up warning them of this happening. This will only happen
when either eventful or geocodingAPI do not return any data.

You can now also search from every activity and the  check for user
location has been expanded a little bit so if the user gives permission
it immediately tries to find a location instead of reporting that no
location is available